requirements.txt
# Big Data Processing
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.2

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Data Analysis
scipy==1.11.4
scikit-learn==1.3.2

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.9
notebook==7.0.6

# Database connectivity
pymongo==4.6.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# Data format support
pyarrow==14.0.1
openpyxl==3.1.2
